backend/requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
environ==1.0
pillow==10.2.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
